build: Bring back dependencies
