+++ Cargo.toml
@@ -4,32 +4,32 @@ version = "0.1.4-rc.44"
 edition = "2021"
 
 [features]
-# completions = ["clap_complete"]
+completions = ["clap_complete"]
 
 [dependencies]
-# clap = { version = "4.5.31", default-features = false, features = ["default", "derive"] }
-# clap_complete = { version = "4.5.46", default-features = false, features = ["default"], optional = true }
-# env_logger = { version = "0.11.6", default-features = false, features = ["default"] }
-# futures-util = { version = "0.3.31", default-features = false, features = ["default"] }
-# ignore = { version = "0.4.23", default-features = false }
-# inference-gateway-sdk = { version = "0.9.0", default-features = false }
-# log = { version = "0.4.26", default-features = false }
-# octocrab = { version = "0.43.0", default-features = false, features = ["default-client", "rustls", "rustls-ring"] }
-# openssl = { version = "0.10.71", default-features = false, features = ["vendored"] }
-# serde = { version = "1.0.218", default-features = false, features = ["derive"] }
-# serde_json = { version = "1.0.139", default-features = false }
-# serde_yaml = { version = "0.9.34", default-features = false }
-# thiserror = { version = "2.0.11", default-features = false }
-# tiktoken-rs = { version = "0.6.0", default-features = false }
-# tokio = { version = "1.43.0", default-features = false, features = ["rt-multi-thread", "tokio-macros"] }
-# uuid = { version = "1.15.1", default-features = false, features = ["v4"] }
+clap = { version = "4.5.31", default-features = false, features = ["default", "derive"] }
+clap_complete = { version = "4.5.46", default-features = false, features = ["default"], optional = true }
+env_logger = { version = "0.11.6", default-features = false, features = ["default"] }
+futures-util = { version = "0.3.31", default-features = false, features = ["default"] }
+ignore = { version = "0.4.23", default-features = false }
+inference-gateway-sdk = { version = "0.9.0", default-features = false }
+log = { version = "0.4.26", default-features = false }
+octocrab = { version = "0.43.0", default-features = false, features = ["default-client", "rustls", "rustls-ring"] }
+openssl = { version = "0.10.71", default-features = false, features = ["vendored"] }
+serde = { version = "1.0.218", default-features = false, features = ["derive"] }
+serde_json = { version = "1.0.139", default-features = false }
+serde_yaml = { version = "0.9.34", default-features = false }
+thiserror = { version = "2.0.11", default-features = false }
+tiktoken-rs = { version = "0.6.0", default-features = false }
+tokio = { version = "1.43.0", default-features = false, features = ["rt-multi-thread", "tokio-macros"] }
+uuid = { version = "1.15.1", default-features = false, features = ["v4"] }
 
 [dev-dependencies]
-# assert_cmd = "2.0.16"
-# assert_fs = "1.1.2"
-# predicates = "3.1.3"
-# serial_test = "3.2.0"
-# tempfile = "3.17.1"
+assert_cmd = "2.0.16"
+assert_fs = "1.1.2"
+predicates = "3.1.3"
+serial_test = "3.2.0"
+tempfile = "3.17.1"
 
 [profile.release]
 lto = true
